WL#15517: [2] HeatWave Support for InnoDB Partitions - Extend MySQL API to support partition (UN)LOAD
This patch modifies the MySQL parser to accept the partition load/unload
commands and also enriches the tables of the RAPID engine with
partitioning information.

a) Partitioning API
====================
The commands we use for loading/unloading partition are the following:
- ALTER TABLE SECONDARY_LOAD PARTITION (p0,p1,..,pn)
- ALTER TABLE SECONDARY_UNLOAD PARTITION (p0,p1,..,pn)

By using these two commands, and following a similar to the familiar
SECONDARY_(UN)LOAD syntax, users can select a set of partitions that
they want to load or unload.

The PARTITION() clause is parsed by MySQL with the updated
SECONDARY_LOAD/SECONDARY_UNLOAD grammar rules in sql_yacc (they are
extended to also carry an optional partition list) and the names are
stored in Table_ref->partition_names.

b) Partitioned secondary tables
===============================
Rapid tables inherit exactly the same partitioning information that
exists in the primary engine.

Change-Id: Idf444787d8b0f9dd0504137a509ec18409e9f567