Bug#35168738: Result diff seen with hypergraph off and on (different number of rows)
When an IN subquery has an always false condition in the WHERE clause,
the hypergraph optimizer may end up performing an IN-to-EXISTS
transformation, but without adding the required HAVING clause to the
transformed query. This causes wrong results.

It is caused by code for creating access paths for empty join results,
reused from the old optimizer, which has the side effect that it sets
the root access path of the JOIN object. Whereas setting the root
access path is fine in the old optimizer, the hypergraph optimizer
expects that the IN-to-EXISTS plan lives in JOIN::m_root_access_path
and the plan without the transformation lives in
JOIN::m_root_access_path_no_in2exists. Since it overwrites the
IN-to-EXISTS plan for the subquery with a non-IN-to-EXISTS plan, it
could end up with a hybrid plan where the outer query block is planned
for IN-to-EXISTS and the subquery is planned without IN-to-EXISTS.

Fixed by changing the code that creates the access paths for empty
join results so that it returns the created access path instead of
modifying the JOIN object. The callers are changed to set
JOIN::m_root_access or JOIN::m_root_access_path_no_in2exists as
appropriate.

Change-Id: I1fd13df11cf71484ad4015e7fc4c19d9201e5459