Bug#35362577 Identify statement boundary while executing stored functions
Currently the MLE SPs lack the notion of 'Statement'. Adding the query-id.
NOTE: Component service interface change only - the actual implementation is to follow in a separate gerrit.
NOTE: The component service is already in the 'mysql-trunk' branch but not yet published to the public.
NOTE: The change is going to be pushed in the 'mysql-trunk' branch, please ignore the gerrit branch displayed.

Change-Id: I405427d4925a838f3bc9fb649fc7bd8d38e2463f