Bug#35067523: heap-use-after-free in sp_head::is_sql when calling stored function
When a stored routine is accessed in a DML operation, sp_head
instance for a routine is created from the data-dictionary instance
of routine. While creating sp_head instance for a routine, instead
of referring to members of data-dictionary members, data members
are copied to sp_head members. But by mistake, only language member
of a sp_head referred to the data-dictionary member.

Data-dictionary instances are stored in the data-dictionary cache
after invoker releases them. The data-dictionary cache elements
are evicted using LRU algorithm. Even routine's data-dictionary
instance is stored in the cache if there is a slot after creating
sp_head instance. And sp_head instance is cached with the THD.
Subsequent access of routines uses the cached sp_head instance.
Since language member of a sp_head is referring to the member of
a data-dictionary instance, accessing language works as expected
till data-dictionary instance is in the cache. If the data-dictionary
instance is evicted from a cache then language member referring to
data-dictionary instance becomes stale. The reported issue
occurred because of the same reason.

To fix this issue, created copy of language string while creating
a sp_head instance for a routine.

Change-Id: Iabc9736af33728a46becdeb624ba0f454e9b95df