Bug#35242734: Signal 11 due to missing UDF file
Description:
With invalid UDF registered, where shared library
for this function could not be loaded, server correctly
reports the following error when trying to use such UDF:
"ERROR 1305 (42000): FUNCTION mysql.udf_test does not exist".
But, if we restart the server and try to call the same invalid UDF again,
server will exit.

Solution:
The fix is to remove UDF from cache when its shared object can not be
loaded on startup.

Change-Id: I42e23cd425d4f03ab138e4b734c73bbbe9beb924