Bug#35167733: Result diff seen with hypergraph off and on (output differs)
The hypergraph optimizer could in some cases omit a required filter on
top of an index merge plan. This could happen if a predicate with a
possible index merge plan was ORed with a conjunction of two
predicates that both had a possible index merge plan.

There was code to detect this in FindIndexRangeScan(), but the check
was indirect. It checked if there were multiple possible index merge
plans for the predicate, which can only happen if there is an AND
between two index merge trees. But there's no guarantee that there are
multiple candidates every time there is an AND between two index merge
trees, so this check does not detect all cases. Specifically,
tree_or() trims down the list of index merge candidates so that it has
only one candidate (it doesn't have to for correctness, but it does it
to reduce the number of plans the range optimizer has to consider). So
an index merge plan for an AND inside an OR might not be correctly
flagged as inexact.

It is fixed by making tree_and() mark the ANDed tree as inexact if
either child has a potential index merge plan, and checking this
property instead in FindIndexRangeScan(). The existing checks for
inexactness of index merges in FindIndexRangeScan() are replaced by
assertions.

Change-Id: Icacdbf39bdaf220b36b54c6afdd5ad17c878bbe4