Bug#35304195: ssl_session_data_print creates world readable files by default
When the ssl_session_data_print mysql client command is being output
to file, the result file is created as world readable.

The solution is to create file with 0600 access permission,
instead of using the existing 0644 default.

Solution is Unix specific, behaviour stays the same on Windows
due to missing group|other permissions concept for this platform
within the MySQL posix wrappers.

Change-Id: I0e9f8040c0f0dd01e46631cab32b56dc9fc451e6