Bug#35129863 Hypergraph: Multi-field indexes ignored in some selectivity estimates
For queries like:

SELECT ... FROM ...
WHERE t1.f1=t2.fX AND ...t1.fN=t2.fY

this commit ensures that if there is a (non-hash) index starting with
some permutation of some subset of {f1..fN}, we will use that index to
calculate the selectivity of those fields (using KEY::records_per_key()).

Prior to this commit, we would only use the index to calculate the
selectivity for the first field in the index. The selectivity of the
remaining fields would be calculated using histograms (if available)
or fallback rules. The selectivities of each individual field would
then be multiplied with each other to get the combined selectivity.
That method will give too low combined selectivity if the values for
the individual fields are correlated, e.g. if
"country='France' AND city='Paris'" is a more likely predicate than
"country='Nepal' AND city='Paris'".

This is implemented by extending the way we represent companion sets.
Companion sets are sets of tables that are inner joined with each other.
In this query:

SELECT ... FROM (t1 JOIN t2 ON ...) LEFT JOIN t3 ON ...

there are thus two companion sets: {t1, t2} and {t3}. Prior to this
fix, companion sets were represented as an integer attribute of the
RelationalExpression class. Now there is a CompanionSet class that
contains the set of sets of fields in equijoin predicates in that
companion set.  For example, if we have:

SELECT ... FROM t1, t2, t3 WHERE t1.x=t2.x AND t2.x=t3.x AND t2.y=t3.y

the CompanionSet will contain the following set of sets:

{{t1.x, t2.x, t3.x}, {t2.y, t3.y}}

Using this information, it is feasible to check if there is a pair of
tables that are joined on fields that form a prefix of some index.
If so, we can use statistics for that index to calculate the selectivity
for each of those fields, rather than just the first one.

Change-Id: I07dcfcef8e6bb975bf9980fe000caa2a945e11c9