Bug#32932636: HYPERGRAPH: ASSERTION `STRATEGY == SUBQUERY_STRATEGY::CANDIDATE_FOR_IN2EXISTS_OR_MAT' FAILED.
An assert failure was with the hypergraph optimizer when executing
certain queries that had an IN subquery nested within another IN
subquery. This happened if the optimizer chose a plan that attempted
to materialize the inner subquery.

The assertion happened because
Item_in_subselect::finalize_materialization_transform() was called
twice on the same subquery, and it is not idempotent. It was called
twice because the Item representing the subquery was referenced twice
the Item tree, and WalkItem() in FinalizeMaterializedSubqueries()
would see it twice.

Fixed by making FinalizeMaterializedSubqueries() skip subqueries that
it has seen before.

Change-Id: Ie2182db04a63fce3fdd55ac02c06c4c406359ccd