Bug#34800383: Hypergraph does not use covering index scans on full table
The hypergraph optimizer does not consider a full index scan on a
table unless it follows an interesting order that can be used to avoid
a sort later. But this is not the only situation in which a full index
scan is cheaper than a full table scan. If the index is covering, the
volume of data that must be read could be considerably smaller with an
index scan than with a table scan.

This patch makes the hypergraph optimizer consider an index scan
whenever there is a covering index on the table.

Change-Id: Ia8de9c289434ec5d118eb3414ae7c26259e4d3ae