Bug #32528899 - LCP_WAIT_CLOSE EMPTY STATE WITH REPEATED "1 BYTE LEFT" ATTEMPTS LOGGED
LCPFragWatchdog (LFW) monitors LCP progress and stops the DB node if no
progress is made for a maximum elapsed time, initialized by
the config variable DB_LCP_SCAN_WATCHDOG_LIMIT = 60000 ms.

From 7.6, Local checkpoint (LCP) completion has a dependency on
one or more Global checkpoints (GCP) completing - if GCPs are not
completing then LCPs cannot complete.

During the phase of operation where a node-local LCP is potentially
waiting for a GCP, the maximum elapsed time is extended to ensure
that it allows for the maximum GCP delay.

GCPs are distributed across all data nodes, and monitored by the GCP
Coordinator in one elected data node.  Due to the distributed nature
of GCP, the GCP Coordinator generally has a loose bound on the maximum
duration of a GCP to avoid falsely attributing communication +
heartbeat problems as GCP problems.  To maintain a minimal bound on
GCP duration, the GCP coordinator's bound is recalculated as data
nodes join and leave a cluster.

Problem:

The fix for 'Bug#27664092 - UNEXPECTED GCP STOP WHEN OTHER NODE SHUTS
DOWN' consolidated the calculation of GCP bound to the GCP Coordinator
at the start of a GCP.  This introduced the regression that only the
LFWs performing LCPs at the Coordinator node receives the
(re)calculated value, and not the LFWs on the other nodes.

LCPs are performed even before a node reaches 'started' phase (101).
Until the (re)calculated values known in time, LFWs may operate with a
tighter configured LCP maximum elapsed time.  If there is a GCP stop
affecting LCPs, non Master LFWs may take escalating action causing a
cluster outage before the GCP Monitor has time to resolve the
situation without an outage.

Fix:
- All GCP participants need to (re)calculate GCP bound at an
  early stage of a node start (start phase 4) and
  whenever a node leaves or joins a cluster, and
  inform their LFW.

- LFW should use it at the start of an LCP phase.

Reviewed by: Frazer Clement <frazer.clement@oracle.com>

Change-Id: I00875f1b2a59bc14668f7ff3ef4e346341f84cab