Bug#34527697: EXPLAIN and EXPLAIN ANALYZE disagree on covering index scans [patch 2/2, fix]
The output from EXPLAIN may not always show that a covering index scan
is covering.

The reason is that it checks TABLE::key_read to determine whether or
not a scan is covering, and this flag does not have to be set until
the scan starts. Exactly when this flag is set varies between type of
scan, and also between which optimizer is used, so it also varies
whether or not the information is shown correctly in EXPLAIN.

Some iterators set the flag in Init(). Since EXPLAIN does not execute
the statement, it does not call Init() on the iterators, and EXPLAIN
may not show that the scan is covering. EXPLAIN ANALYZE does execute
the statement, so it may sometimes get it right. Other times it is
wrong for EXPLAIN ANALYZE too, because some iterators reset the flag
once the scan is done (for example, SortingIterator sets
TABLE::key_read to false on all input tables once it is done reading
them).

This patch fixes it by using a different way of detecting that an
index scan or index lookup is covering. EXPLAIN now checks if the
index is marked as covering in TABLE::covering_keys and that covering
scans have not been disabled on the table by setting the
TABLE::no_keyread flag. This is the same check as the execution code
performs to find out if a scan should be covering.

Note that the patch removes the code that tries to figure out if a
dynamic range scan is covering. Since a dynamic range scan is
re-planned for each row in the outer table, it does not make much
sense to say that a scan is covering or not. In the same execution, it
could be covering for one row and not covering for another row. And in
any case we don't know what it will be when we EXPLAIN the statement,
since the plan is not decided before execution has started.

Change-Id: I1ff96a0c9e997c54dae28736097a1710894cb7ad