Fix hard-coded timestamps in main.derived_condition_pushdown MTR test
Two hard-coded timestamps were added to this test in
https://github.com/mysql/mysql-server/commit/22f6ef90cae8#diff-d7e1b8c7138b94f92fbd555c44e7facc027cb50217f59f51729444ee7c3b37a4R1950

1. '2022-05-06 16:49:45' (assumed by the test's result file to be `< NOW()`)
2. '2023-05-06 16:49:45' (assumed by the test's result file to be `>= NOW()`)

Because the second of these timestamps is now in the past, the test is now
failing due to result mismatch.

Tests should not be written to include hard-coded assumptions about the
local time in their execution environment; these are essentially guaranteed
to get out-of-date and cause failures.

This test can very easily be rewritten to choose past/future timestamps
relative to the time at which it is run,
`DATE_ADD(NOW(), INTERVAL [+-] 1 DAY)`.

Note that Ubuntu's build environment had to disable this test on 2022-05-09,
likely due to encountering this problem:

    https://bugs.launchpad.net/ubuntu/+source/mysql-8.0/8.0.33-0ubuntu2

All new code of the whole pull request, including one or several files that are
either new files or modified ones, are contributed under the BSD-new license. I
am contributing on behalf of my employer Amazon Web Services, Inc.