Bug#35377192: Failure when executing a subquery inside a stored procedure               with optimizer trace
This problem may occur when executing a stored procedure with an
IF statement or similar that contains a subquery, when optimizer trace
is enabled. On second execution, MySQL attempts to trace the expression
containing the subquery, however this occurs before the expression
is properly bound with data from tables, and the print() function
that traces the statement may hit a NULL pointer.

The solution is to ensure that table data are properly bound to the
statement before calling the trace function. This is also what happens
in optimization of regular SQL statements, where we trace properly
bound statements inside JOIN::optimize().

We also do a small refactoring by moving the execution code for the
expression into a separate function, and also move all cleanup code
into a scope guard function, thus making it no longer necessary to
keep track of execution status in a local variable.

Notice that the call to bind_fields() is strictly not required for the
first execution, but it also does no harm.

Change-Id: Ia6f5b78ef1ae4b641058a211f0a93c65f08096a6