BUG#35373030: Relax read lock acquire on GR network and message metrics
It was observed that any query to get server status variables will
fetch all status variables.
Example, the below queries will fetch all status variables,
including the ones named `Gr_*` and only then filter to only show
the requested one.
  SHOW STATUS LIKE 'Uptime';
  SELECT * FROM performance_schema.global_status WHERE VARIABLE_NAME='Uptime';
Some `Gr_*` status variables do need to acquire a read lock, if a
exclusive operation - like `START GROUP_REPLICATION` - is ongoing,
that will make `SHOW STATUS` to wait until that operation is
complete.

To avoid delay `SHOW STATUS`, the freshness of the value of some
`Gr_*` status variables is relaxed, that is, there is a attempt to
acquire the read lock, if it is successful the latest value is
returned, if the lock acquire is not possible, a cached value is
returned. On both cases, the `SHOW STATUS` is not delayed.

Change-Id: I1b2a0b9366d0fb6e6351ac466f105a7e076f8812