Bug#33824383: hypergraph: Unnecessary materialization after aggregation [patch 1/2, refactor, noclose]
Initial refactoring of the hypergraph_optimizer-t unit test.

Reduce the amount of boilerplate needed to create an index which
supports ordered scans and range scans. A new helper function,
CreateOrderedIndex(), is added. It creates an index over a set of
columns and instructs Mock_HANDLER to return index flags that indicate
support for ordered scans and range scans. Existing test cases are
updated to use the new helper function.

Change-Id: I112e36959421a102bb6f3b7d7a2a3561723a5498