Bug#35082046: AccessPath* {anonymous}::CostingReceiver::ProposeAccessPath Assertion `false' Bug#35096962: Assertion : `false' in (anonymous namespace)::CostingReceiver::ProposeAccessPath
Post-push fix for ASAN failure in the unit test case. It was a stack
buffer overflow in SortNodes() which happened because it was not
prepared for seeing predicates which accessed pseudo tables, and it
ended up reading outside of the array that remapped old node indexes
to new node indexes.

Fixed by not trying to remap pseudo tables.

Change-Id: I0ebd3af19a514e3ad4c7072aca476dc683180622