Bug#35385801 buf_page_free_stale_during_write() misses os_event_set(buf_pool->no_flush[i])
When flushing dirty pages to tablespaces we use BUF_FLUSH_LRU,
BUF_FLUSH_LIST or BUF_FLUSH_SINGLE_PAGE flush_type, and for each of them
we track:

- n_flush[flush_type]
    the number of pages for which write operation was started,
    but has not completed yet

- init_flush[flush_type]
    a boolean indicator that we are preparing a batch of several pages
    to write and for some of them the writes were already started (and
    for some perhaps even finished already)

- no_flush[flush_type]
    an event which is in "set" state iff there is no ongoing write and
    no batch being initialized at the moment:
    n_flush[flush_type] == 0 && !init_flush[flush_type]

Note that n_flush[flush_type] may drop to zero even while
init_flush[flush_type] is still true, i.e. there are still some writes
to be issued. This is why it is often useful to wait for the combination
of the two factors, which is what no_flush[flush_type] tracks.

Also, for BUF_FLUSH_SINGLE_PAGE we never use batching (as the name
suggest) so the init_flush[BUF_FLUSH_SINGLE_PAGE] is always false.

To know when to set() and reset() the no_flush[flush_type] event we keep
track of the moments when the n_flush[flush_type] and
init_flush[flush_type] change values to see if the change impacts the
truth value of n_flush[flush_type] == 0 && !init_flush[flush_type].

The buf_page_free_stale_during_write() has a bug, though, in that it
decrements n_flush[flush_type] without checking if it has dropped to
zero, in which case it might mean no_flush[flush_type] should be set (if
additionally init_flush[flush_type] is already false).

This bug fix adds the missing call to os_event_set().
It also cleans up the code a bit by:
- introducing bool is_flushing(flush_type) predicate explicitly
- ensuring no_flush[flush_type] is initially in "set" state according to
  its definition
- ensuring that the state of no_flush[flush_type] always matches the
  is_flushing(flush_type) by introducing a
  change_flush_state(flush_type, foo) helper which checks if
  the value of is_flushing(flush_type) has changed when executing foo()
  and reflects the change on the no_flush[flush_type] accordingly
- above in turn makes the buf_flush_wait_LRU_batch_end() unnecessary
- making sure all types of flush are correctly and uniformly handled in
  various places, in particular adding BUF_FLUSH_SINGLE_PAGE to BP stats
- removing a declaration for undefined buf_flush_wait_batch_end_wait_only

Change-Id: Ie0029de1f5e774e019cf1da7323fa6dfe1a6755b