Bug#33824383: hypergraph: Unnecessary materialization after aggregation [patch 2/2, fix]
When adding sort paths for ORDER BY or DISTINCT, the hypergraph
optimizer adds a materialization step before the sorting if there are
aggregates to sort. This step is needed by Filesort so that it has a
record buffer to write the sorted aggregated values to.

If the sort is elided because the result after aggregation already
follows the desired ordering, the materialization step is unnecessary.
The hypergraph optimizer still adds materialization after aggregation
if there is an ORDER BY or DISTINCT clause.

This patch changes ApplyDistinctAndOrder() so that it adds a
materialization step only if it actually adds a sort node.

Change-Id: Ia75bc40c3f6356604f71add2ca9d8e28abea7810