Bug#35382014: Mysqld crash: Assertion `item_name.is_set()' failed in sql/item.cc
Code that prints the GROUP BY clause for inclusion in the output from
EXPLAIN FORMAT=JSON hit an assertion in certain situations when
running with the hypergraph optimizer.

The code that hit the assertion tried to print the alias instead of
the full expression, if an alias had been used in the GROUP BY clause
of the query originally. It failed because of the item_name of the
Item representing the expression was not set.

Query_term::print_order() calls Item::print_for_order() on each GROUP
BY item. These items do have the name set, but
Query_term::print_order() unwraps any rollup group wrappers first, so
it might be called on a child of the GROUP BY item. The child of a
rollup group wrapper usually has the same item_name as the wrapper,
since the wrapper gets its name from the child in the first place.
However, in the query that hits the assertion, the child of the rollup
group wrapper has been replaced by an Item_cache which has not been
given a name.

Fixed by changing the resolving of ORDER BY and GROUP BY elements to
remember not only whether an alias was used, but also what the alias
was. Then we don't need to look up the alias from the item later, and
it is no longer a problem that the expression is transformed during
optimization.

Change-Id: I4c156869f817e6db4f30a2dd4955dbf7696a0c6c