Bug#35327006 Stabilize the weekly runs [noclose]
Prevent rpl.rpl_deprecations from running with transaction_write_set_extraction
OFF or replica_parallel_type=LOGICAL_CLOCK as this generates extra
deprecation warnings that make the test fail.

Introduce a skip file that tests if one or more options are specified
on the MTR command line and use this to replace the sequence of
different have_*.inc invocations in the test.

Change-Id: I7a5fff8d0b80d4fcf531cf6d25996b2a1ac564c6