Fix the CPPCHECK mode for the latest VS2022 (1936).
When compiling table_histograms with VS2022 with MSVC_CPPCHECK=1 I get:
33>------ Build started: Project: sql_main, Configuration: Debug x64 ------
33>table_histograms.cc
33>C:\Users\gkodinov\dev\mysql-trunk\sql\histograms\table_histograms.cc(146,1): error C2220: the following warning is treated as an error
33>C:\Users\gkodinov\dev\mysql-trunk\sql\histograms\table_histograms.cc(142): warning C28020: The expression '0<=_Param_(1)&&_Param_(1)<=16-1' is not true at this call.: Lines: 132, 133, 134, 135, 133, 134, 135, 133, 138, 142

This is a VS2022 issue that I've reported as https://developercommunity.visualstudio.com/t/False-positive-warning-C28020/10362798

Until the underlying issue is fixed we apply the assert() workaround.

Also suppressed one extra warning that is not very well applied to
various code fragments that are otherwise valid.
And added a suppression on warnings in files using winSDK headers that
generate /analyze errors.

Change-Id: I665abc2fd4c6996ba10005f391627466e58be43d