BUG#35361668: Fix error reporting when secondary load thread is killed
If an ongoing secondary load DDL is interrupted by the client using
CTRL+C or KILL QUERY command, then based on the stage at which the DDL
was before the interruption, it may throw different errors to the user.

For example, if DDL is aborted before secondary engine operation is
completed, then it returns ER_QUERY_INTERRUPTED; However, if the DDL
is aborted after secondary engine action is finished but before it
reaches commit phase, then it returns ER_RUN_HOOK_ERROR; Additionally,
for latter case, an error message with severity 'ERROR' is added to
the mysqld error log saying 'transaction commit failed'. This error
message causes confusion to MTR tests as they don't expect an 'error'
message in case of intentional interruption.

In this commit, we change the message being reported in the errorlog.
If thread was killed intentionally, we log a message with severity
'SYSTEM' saying 'query aborted before commit'.
Otherwise, if commit failed for some other reasons, we log an error
message with severity 'ERROR' saying 'transaction commit failed'.

This commit is not changing the error message that the mysql client
(and thus the user) gets when the DDL fails or is aborted by user.

When LOAD aborts from secondary engine, it gets ER_QUERY_INTERRUPTED;
When LOAD aborts from transaction commit, it gets ER_RUN_HOOK_ERROR;

New test is added to validate the behavior.

Change-Id: Ifdf04b4c071a7cf851098ffdb6480bb4031d22ce