Bug#35362577 Identify statement boundary while executing stored functions
Currently the MLE SPs lack the notion of 'Statement'. As per the discussion with the Component Service team, we are adding the Statement Handle to facilitate a future Component Service.
NOTE: Component service interface change only - the actual implementation is to follow in a separate gerrit.
NOTE: The affected component service is already in the 'mysql-trunk' branch but not yet published to the public.

Change-Id: Iee0a97440bd7005f2d4da1a1b8b87ef49adac34c