WL#15002 NDB: Explicit epoch mem_root usage [binlog index rows 2]
Extract class maintaining the list of server_id/epoch and related stats
handled during an epoch. This avoids the repeated allocation of memory
from MEM_ROOT in favour of using a small piece of memory reused for
every epoch. As part of this the data structure is reduced by making:
1) the counter for schemaops to be "global" during the handled epoch.
2) the fixed parameters which are written to every row in
   mysql.ndb_binlog_index at commit are passed as direct
   arguments to write function rather than indirectly via the list.
In addition improve function comments and rename the mystery "flag"
argument to "orig_epoch".

Change-Id: I4b1387e2799a4bab6c108e5ae44d21e6c438d13f