Bug#33824300: hypergraph: Wrong results from hash join with long string types [patch 3/3, fix]
If a hash join is performed using an equijoin condition between string
columns whose maximum weight string length exceeds 1024 bytes, the
full column value is not stored in the hash table. A hash is stored
instead. Because of this, the hash join may return too many rows if
there are hash collisions.

The old optimizer removes the extra rows by evaluating the equijoin
condition together with the non-equijoin conditions after the hash
join, when needed. This is done either in a separate filter after the
hash join (for inner joins), or as a so-called "extra condition"
attached to the hash join (for outer joins).

The hypergraph optimizer does not currently evaluate such join
conditions after the join, and could therefore return too many rows.
This patch makes the hypergraph optimizer, too, evaluate the condition
after the hash join. It does this by adding the equijoin conditions on
long string columns as "extra conditions" when creating the
HashJoinIterator.

Note that, contrary to the old optimizer, it adds these conditions to
the "extra conditions" for both inner and outer join. It does not
create a separate filtering step for inner joins. This does not make
any difference for inner join. The only difference is whether the
filtering happens inside the hash join iterator or right after the
hash join iterator. (For outer join it does matter, since a filter
would not add null-complemented rows as needed.) This is consistent
with how non-equijoin conditions are currently handled in the
hypergraph optimizer, and it's an existing difference between the two
optimizers.

Change-Id: Id4e9e787ba7d6b302d58eb6e60b15292139a0dbe