BUG#34398622 gr_suspect_member_resumes_after_crash_mysql bug failed
Problem
========================
gr_suspect_member_resumes_after_crash_mysql is failing sporadically
on our continuous integration platform on SLL runs

Analysis
========================
MySQL connections used in XCom are not fully controlled by whom who
creates them but by:
- The Network Connection Manager that lies in XCom;
- The server when it goes into a stop.

In this case, it means that the GR plugin might be stopping, but
the server is still accepting connections that are to be delegated
to GCS.

In a group with 2 nodes (A,B), a Node B XCom client that creates a
connection to join a group is not aware that the new connection is not
making any progress on Node A server-side, thus it blocks on a con_read
that will never return with success.

This lack of return derived from the fact that:
- Plugin was stopping and it would never return anything on that
  connection;
- Server side plugin would get stuck trying to acquire a read lock
  on a resource that is not available when the server is either stopping
  or starting.

Adding to that, connections that are received when the plugin is
stopping might end up in a limbo, because they were accepted but never
registered on the Network Connection Manager from XCom.

Fix
========================

The fix has several fronts, all of them on the server side:
- We changed the sql_parse.cc code to return client errors during the
  connection delegation itself. This allows for a client to detect
  any of the errors described above, fail-fast and retry the connection.

- Change the lock acquisition to a "try lock", returning an error if
  we can't obtain a lock to delegate the connection, when the plugin
  is executing an exclusive operation on gcs_operations.cc.

- Changed the wait loop for a delegated server connection to check if
  the plugin is still enabled. If the plugin is tear down, those
  connections are terminated.

Change-Id: I0528e1df9adcd985181a233e226b6352af4eae7c