Bug#35393351: MLE stored programs do not take correct sql-mode into account while executing Bug#35383837: Modify MTR tests for float/double to create functions with non-strict sql-mode
Problem:
While executing a routine, sql_mode and the database used while creating
a routine must be used. But for the MLE routine, sql_mode and the database
of a session were used.

Also, in the existing float and double-type conversion tests,
we are creating MLE stored programs using strict mode and using the same
functions to test the non-strict mode behavior. However, we will not be
able to test the non-strict sql-mode if we do not create the functions
with non-strict mode as this behavior will be enforced by the fix
for bug#35393351.

Fix:
Make sure the MLE stored programs take into account the same sql_mode,
database during execution as the one during creation.
Modify the test files to create the stored programs again so that we
can test the non-strict sql-mode behavior for type conversion tests.

Change-Id: I36ab29838b0485169c365dae010ec08619ebae83