Bug#33824300: hypergraph: Wrong results from hash join with long string types [patch 1/3, move test, noclose]
Move the contents of hash_join.test into a new file hash_join.inc, and
make hash_join.test source the new file. This is done in preparation
for making the test run in two configurations; with the old optimizer
and with the hypergraph optimizer. It is done in a separate commit,
without any changes to the contents of the file, to make it easier for
git to track the history across the renaming.

Change-Id: I4387099bc202e20fdc2aba34eceeeb97646ecf2b