WL#15497: Add statement type "DDL" to transaction tracking facility Bug#34664197: `group_replication_set_as_primary()` does not wait for ongoing ALTER TABLE
The MySQL server keeps track of certain attributes for
each connection, including whether a transaction is ongoing,
whether it has work attached, whether the current statement
is DML, whether tables are locked, and so on.
This information can be queried within the server to
inspect another session's state and attributes;
some of it can additionally be reported to clients
who have subscribed to such information.

An example case of the latter mechanic is a load-balancer
that is positioned between clients and servers and that
needs to know when a connection can safely be moved to a
different server (e.g. while the session is not in the
middle of a transaction that would then be rolled back).

However, while we currently track whether a running
statement is DML, we do not provide similar information
for ongoing DDL statements. This WL aims to add this
support in order to significantly improve the group
replication and Heatwave user experience.

It is required for WL#10378: "Group Replication: group
single/multi primary mode change and primary election"
to function as expected:
`group_replication_set_as_primary()` must wait
for all ongoing transactions on the previous
primary to complete before electing a new primary.
DDL such as ALTER TABLE is not recognized, and
thus the MySQL does not wait for it to complete
before electing a new primary. (Bug#34664197).

Notes

REFACTORED CLASSES AND STATEMENTS; TEST CASES

Sql_cmd and sub-classes used to offer is_dml(),
which returned `true` once a statement had been
identified as DML, and `false` if it had been
identified as "not DML" (DDL or otherwise), or
had not been identified as anything yet. This
method is replaced with sql_cmd_type() which
makes the distinction between the states clearer.

As before this patch, the classification in the
code and the SQL standard still disagree on a
few commands, but now we have a mechanism
in place that we let us recategorize commands,
add new ones, and add new categories.

We also refactor the inheritence so that commands
inheriting from Sql_cmd_ddl_table correctly
inherit from Sql_cmd_ddl as well. The following are
affected and now have dedicated test cases:

- OPTIMIZE TABLE
- TRUNCATE TABLE   (inheritance adjusted explicitly)
- ANALYZE  TABLE
- ALTER    TABLE
- CHECK    TABLE
- CREATE   TABLE
- DROP     TABLE   (Sql_cmd_drop_table added)
- CREATE   INDEX
- DROP     INDEX

This leads to the above reporting SQL_CMD_DML
via sql_cmd_type().

DML / RUN_BEFORE_DML_HOOK()

In run_before_dml_hook(), we now only flag TX_STMT_DML
for statements that do not identify as SQL_CMD_DDL
(as some statements, such as CREATE TABLE ...
AS SELECT ..., go through both DML and DDL code paths).

DDL / HANDLER

As the analogue to the above, DDL will generally pass
through handler::mark_trx_read_write(), but not
everything that passes through there is DDL. Thus, when
a statement passes through there and also identifies as
DDL (via thd->lex->m_sql_cmd->sql_cmd_type()), we mark
the statement as DDL in the tracker.

SESSION TRACKER "TRANSACTIONS"

The transaction tracker subystem of the session
tracker engine already flagged DML statements
as such. It did not do so for DDL, partially
because there was no simple interface for
detecting that (which this changeset now adds),
and partially because commands that generally
would auto-commit could not become part of a
multi-statement transaction anyway. As such,
reporting DDL only after the statement completes
(as the session tracker does) seemed of limited
value. For this reason, keeping the transaction
tracker interface stable was considered more valuable
than after-the-fact reporting. (This could however
be easily changed in subsequent instalments now,
building on the groundwork added with this changeset.)

Other threads within the server however can inspect
this tracker's findings, now including whether or
not the statement was DDL.

GET_ONGOING_SERVER_TRANSACTIONS

This is a component function in the
mysql_ongoing_transactions_query API offered
by the MySQL server. It returns to its callers
a count of sessions in the MySQL that are
considered as having a transaction underway.
It also reports the thread-IDs of those sessions.

This mechanism is used by Group Replication
to determine whether it is safe to change
over to a different PRIMARY. As this entails
setting the old primary `super_read_only`,
it should only be done while no transactions
are ongoing.

While this API as such has not changed, it
will now report more complete information
(e.g. it will include DDL/single statement
transactions). This, in a manner of speaking,
is this WL's "endgame", its raison d'etre.

For this purpose, the mechanism queries the state
of each THD's `transaction` session-tracker (if
present).

We also consider the Sql_cmd's sql_command_flags
(such as CF_CHANGES_DATA, CF_IMPLICIT_COMMIT_BEGIN,
CF_IMPLICIT_COMMIT_END) to this end where available.

Change-Id: Ia90487a2404690976dcd6ac910838960c0966dbe