Bug#35414464 Single TLS connection triggers 2 sessions cache hits
Patch for "Bug#35339886 Use compliant primes for DH and curves for ECDH"
limited the list of the supported curves groups to the 3 elements: secp256r1,
secp521r1, secp384r1. This causes the client (Router in this case) list
to be incomaptible as the first group on the Router's list is
ecdh_x25519 which the server does not support now. As a result the handshake
procedure is restarted and the Router uses supported secp256r1 the second time.

This patch limits the curves groups on the Router side to the same 3
elements to match the server's list and avoid the additional TLS handshake
round-trip.

Change-Id: Ifd3115ed2118bc4049dbc8e06664f2634785d674